+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2022-10-31
 1. GRL [ADD] Allow the user to get all the product_catalogs
 2. GRL [ADD] Allow the user to get a single product_catalog
+3. GRL [ADD] Allow the user to create new product_catalogs
 
 ## 202-10-24
 1. GRL [ADD] Allow the user to manage labels